requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-multipart==0.0.5
jsonschema==4.17.3
PyGithub==1.57
black==22.8.0
pygelf==0.4.2
heatshrink2==0.13.0
pillow==10.3.0
